Drop dead code. Move distribution zone exception from code module. Minor fixes.
