+++ modules/catalog/src/main/java/org/apache/ignite/internal/catalog/CatalogParamsValidationUtils.java
@@ -18,7 +18,6 @@
 package org.apache.ignite.internal.catalog;
 
 import static org.apache.ignite.internal.lang.IgniteStringFormatter.format;
-import static org.apache.ignite.lang.ErrorGroups.Catalog.VALIDATION_ERR;
 
 import com.jayway.jsonpath.InvalidPathException;
 import com.jayway.jsonpath.JsonPath;
@@ -54,7 +53,6 @@ public class CatalogParamsValidationUtils {
 
         if (value < min || (max != null && value > max)) {
             throw new CatalogValidationException(
-                    VALIDATION_ERR,
                     "{}: [value={}, min={}" + (max == null ? ']' : ", max={}]"),
                     errorPrefix, value, min, max
             );
@@ -71,7 +69,6 @@ public class CatalogParamsValidationUtils {
     ) {
         if (autoAdjust != null && (scaleUp != null || scaleDown != null)) {
             throw new CatalogValidationException(
-                    VALIDATION_ERR,
                     "Not compatible parameters [dataNodesAutoAdjust={}, dataNodesAutoAdjustScaleUp={}, dataNodesAutoAdjustScaleDown={}]",
                     autoAdjust, scaleUp, scaleDown
             );
@@ -92,7 +89,6 @@ public class CatalogParamsValidationUtils {
             String error = e.getMessage() == null ? "Unknown JsonPath compilation error." : e.getMessage();
 
             throw new CatalogValidationException(
-                    VALIDATION_ERR,
                     "Invalid filter: [value={}, error={}]",
                     e,
                     filter, error
